fix insight & chart tsc deps
