+++ ui/chart/tsconfig.json
@@ -8,6 +8,6 @@
     "declaration": true,
     "composite": true
   },
-  "references": [{ "path": "../common/tsconfig.json" }],
+  "references": [{ "path": "../common/tsconfig.json" }, { "path": "../ceval/tsconfig.json" }],
   "isolatedModules": true
 }
