+++ ui/insight/tsconfig.json
@@ -7,5 +7,5 @@
     "composite": true
   },
   "isolatedModules": true,
-  "references": [{ "path": "../common/tsconfig.json" }]
+  "references": [{ "path": "../common/tsconfig.json" }, { "path": "../chart/tsconfig.json" }]
 }
